+++ test/test_basic/test_main.cpp
@@ -8,17 +8,19 @@ void test_api_compiles() {
     ESPTimerConfig cfg;
     timer.init(cfg);
 
-    auto id1 = timer.setTimeout([]() {}, 10);
+    auto id1 = timer.setTimeout([]() {}, 1000);
     auto id2 = timer.setInterval([]() {}, 20);
     auto id3 = timer.setSecCounter([](int) {}, 1000);
     auto id4 = timer.setMsCounter([](uint32_t) {}, 100);
     auto id5 = timer.setMinCounter([](int) {}, 60000);
+    auto id6 = timer.setTimeout([]() {}, 1000);
 
     TEST_ASSERT_TRUE(id1 > 0);
     TEST_ASSERT_TRUE(id2 > 0);
     TEST_ASSERT_TRUE(id3 > 0);
     TEST_ASSERT_TRUE(id4 > 0);
     TEST_ASSERT_TRUE(id5 > 0);
+    TEST_ASSERT_TRUE(id6 > 0);
 
     // Pause then resume; both should return true if found and state changed
     TEST_ASSERT_TRUE(timer.pauseInterval(id2));
@@ -29,6 +31,9 @@ void test_api_compiles() {
     TEST_ASSERT_FALSE(running);
     running = timer.toggleRunStatusInterval(id2);
     TEST_ASSERT_TRUE(running);
+
+    TEST_ASSERT_TRUE(timer.clearTimeout(id1));
+    TEST_ASSERT_TRUE(timer.clearTimer(id6));
 
     // Clear should return true once
     TEST_ASSERT_TRUE(timer.clearInterval(id2));